new idea for css layout for github pages
see the css for comments about the methods.

i left colored borders so you can see where things land,
get rid of them when you need them.

works as expected in FF, Safari, IE6 and IE7.  IE8 does
not seem to respect the min-width on the #wrapper but
it is not a big deal.